requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==2.2
django-allauth==0.40.0
django-rest-auth==0.9.5
django-shortcuts==1.6
djangorestframework==3.10.3
flake8==4.0.1
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
pycodestyle==2.8.0
pyflakes==2.4.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.1.2
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
